requirements.txt
click>=8.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
rich>=10.0.0  # For better CLI output
pypandoc>=1.11 